requirements.txt
einops==0.8.1
numpy==2.2.5
opencv_contrib_python==4.10.0.84
opencv_python==4.8.1.78
PyYAML==6.0.2
scipy==1.15.2
scikit-image>=0.20.0
timm==0.9.12
torch>=2.2.0
tqdm==4.66.1
jax>=0.4.0
jaxlib>=0.4.0
flax>=0.7.0
optax>=0.1.0
wandb>=0.15.0
